# src/main/resources/babymeal-api-spec.yaml
openapi: 3.0.3
info:
  description: 'This is the api spec for UrFuture project.'
  version: '1.0.0'
  title: 'Swagger UrFuture'
  contact:
    email: 'babiqit@gmail.com'
...
servers:
  - url: http://localhost:8080/api/v0
    description: Local machine
paths:
...
  /ingredients:
    post:
      tags:
      - ingredient
      summary: 'Create a new ingredient'
      operationId: createIngredient
      requestBody:
        description: 'The prepared ingredient request to be created.'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngredientModel'
      responses:
        '201':
          description: 'The ingredient is created.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngredientModel'
        '300':
          description: 'Moved: The resource has moved.'
        '400':
          description: 'Bad Request: The request could not be understood by the server due to malformed syntax. Something like Domain validation errors, missing data, invalid input, etc.'
        '401':
          description: 'Unauthorized: The caller is not authorized.'
        '403':
          description: 'Forbidden: The caller has no permission.'
        '404':
          description: 'Not Found: The server has not found anything matching the Request-URI.'
        '408':
          description: 'Timeout: The request timeout.'
    get:
      tags:
      - ingredient
      summary: 'Find all Ingredients'
      operationId: getAllIngredients
      responses:
        '200':
          description: 'All Ingredients'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IngredientModel'
        '300':
          description: 'Moved: The resource has moved.'
        '400':
          description: 'Bad Request: The request could not be understood by the server due to malformed syntax. Something like Domain validation errors, missing data, invalid input, etc.'
        '401':
          description: 'Unauthorized: The caller is not authorized.'
        '403':
          description: 'Forbidden: The caller has no permission.'
        '404':
          description: 'Not Found: The server has not found anything matching the Request-URI.'
  /ingredients/{ingredientId}:
    get:
      tags:
      - ingredient
      summary: 'Finds ingredient by id'
      operationId: findIngredientById
      parameters:
        - name: ingredientId
          in: path
          description: 'ID of ingredient to return'
          required: true
          schema:
            type: string
            maxLength: 64
            description: 'The ingredient identifier (UUID).'
            example: '7f7acbad-8e3b-470f-a3d4-e750360e3964'
      responses:
        '200':
          description: 'The ingredient'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngredientModel'
        '300':
          description: 'Moved: The resource has moved.'
        '400':
          description: 'Bad Request: The request could not be understood by the server due to malformed syntax. Something like Domain validation errors, missing data, invalid input, etc.'
        '401':
          description: 'Unauthorized: The caller is not authorized.'
        '403':
          description: 'Forbidden: The caller has no permission.'
        '404':
          description: 'Not Found: The server has not found anything matching the Request-URI.'
    put:
      tags:
      - ingredient
      summary: 'Update an existing Ingredient'
      operationId: updateIngredient
      parameters:
        - name: ingredientId
          in: path
          description: 'ID of ingredient to return'
          required: true
          schema:
            type: string
            maxLength: 64
            description: 'The ingredient identifier (UUID).'
            example: '7f7acbad-8e3b-470f-a3d4-e750360e3964'
      requestBody:
        description: 'The prepared ingredient request to to be updated.'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngredientModel'
      responses:
        '200':
          description: 'The ingredient is updated.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngredientModel'
        '300':
          description: 'Moved: The resource has moved.'
        '400':
          description: 'Bad Request: The request could not be understood by the server due to malformed syntax. Something like Domain validation errors, missing data, invalid input, etc.'
        '401':
          description: 'Unauthorized: The caller is not authorized.'
        '403':
          description: 'Forbidden: The caller has no permission.'
        '404':
          description: 'Not Found: The server has not found anything matching the Request-URI.'
        '408':
          description: 'Timeout: The request timeout.'
        '409':
          description: 'Conflict: Indicates that the request could not be processed because of conflict in the current state of the resource.'
    delete:
      tags:
        - ingredient
      summary: 'Delete all ingredient by profile id'
      parameters:
        - name: ingredientId
          in: path
          description: 'ID of ingredient to return'
          required: true
          schema:
            type: string
            maxLength: 64
            description: 'The ingredient identifier (UUID).'
            example: '7f7acbad-8e3b-470f-a3d4-e750360e3964'
      operationId: deleteIngredient
      responses:
        '204':
          description: 'The ingredient has been sucessfully deleted.'
        '300':
          description: 'Moved: The resource has moved.'
        '400':
          description: 'Bad Request: The request could not be understood by the server due to malformed syntax. Something like Domain validation errors, missing data, invalid input, etc.'
        '401':
          description: 'Unauthorized: The caller is not authorized.'
        '403':
          description: 'Forbidden: The caller has no permission.'
        '404':
          description: 'Not Found: The server has not found anything matching the Request-URI.'
        '408':
          description: 'Timeout: The request timeout.'
  /ages:
    post:
      tags:
      - age
      summary: 'Create a new age'
      operationId: createAge
      requestBody:
        description: 'The prepared age request to be created.'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgeModel'
      responses:
        '201':
          description: 'The age is created.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgeModel'
        '300':
          description: 'Moved: The resource has moved.'
        '400':
          description: 'Bad Request: The request could not be understood by the server due to malformed syntax. Something like Domain validation errors, missing data, invalid input, etc.'
        '401':
          description: 'Unauthorized: The caller is not authorized.'
        '403':
          description: 'Forbidden: The caller has no permission.'
        '404':
          description: 'Not Found: The server has not found anything matching the Request-URI.'
        '408':
          description: 'Timeout: The request timeout.'
    get:
      tags:
      - age
      summary: 'Find all ages'
      operationId: getAllAges
      responses:
        '200':
          description: 'All ages'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AgeModel'
        '300':
          description: 'Moved: The resource has moved.'
        '400':
          description: 'Bad Request: The request could not be understood by the server due to malformed syntax. Something like Domain validation errors, missing data, invalid input, etc.'
        '401':
          description: 'Unauthorized: The caller is not authorized.'
        '403':
          description: 'Forbidden: The caller has no permission.'
        '404':
          description: 'Not Found: The server has not found anything matching the Request-URI.'
  /ages/{ageId}:
    get:
      tags:
      - age
      summary: 'Finds age by id'
      operationId: findAgeById
      parameters:
        - name: ageId
          in: path
          description: 'ID of age to return'
          required: true
          schema:
            type: string
            maxLength: 64
            description: 'The age identifier (UUID).'
            example: '7f7acbad-8e3b-470f-a3d4-e750360e3964'
      responses:
        '200':
          description: 'The age'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgeModel'
        '300':
          description: 'Moved: The resource has moved.'
        '400':
          description: 'Bad Request: The request could not be understood by the server due to malformed syntax. Something like Domain validation errors, missing data, invalid input, etc.'
        '401':
          description: 'Unauthorized: The caller is not authorized.'
        '403':
          description: 'Forbidden: The caller has no permission.'
        '404':
          description: 'Not Found: The server has not found anything matching the Request-URI.'
    put:
      tags:
      - age
      summary: 'Update an existing age'
      operationId: updateAge
      parameters:
        - name: ageId
          in: path
          description: 'ID of age to return'
          required: true
          schema:
            type: string
            maxLength: 64
            description: 'The age identifier (UUID).'
            example: '7f7acbad-8e3b-470f-a3d4-e750360e3964'
      requestBody:
        description: 'The prepared age request to to be updated.'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgeModel'
      responses:
        '200':
          description: 'The age is updated.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgeModel'
        '300':
          description: 'Moved: The resource has moved.'
        '400':
          description: 'Bad Request: The request could not be understood by the server due to malformed syntax. Something like Domain validation errors, missing data, invalid input, etc.'
        '401':
          description: 'Unauthorized: The caller is not authorized.'
        '403':
          description: 'Forbidden: The caller has no permission.'
        '404':
          description: 'Not Found: The server has not found anything matching the Request-URI.'
        '408':
          description: 'Timeout: The request timeout.'
        '409':
          description: 'Conflict: Indicates that the request could not be processed because of conflict in the current state of the resource.'
    delete:
      tags:
        - age
      summary: 'Delete all age by profile id'
      parameters:
        - name: ageId
          in: path
          description: 'ID of age to return'
          required: true
          schema:
            type: string
            maxLength: 64
            description: 'The age identifier (UUID).'
            example: '7f7acbad-8e3b-470f-a3d4-e750360e3964'
      operationId: deleteAge
      responses:
        '204':
          description: 'The age has been sucessfully deleted.'
        '300':
          description: 'Moved: The resource has moved.'
        '400':
          description: 'Bad Request: The request could not be understood by the server due to malformed syntax. Something like Domain validation errors, missing data, invalid input, etc.'
        '401':
          description: 'Unauthorized: The caller is not authorized.'
        '403':
          description: 'Forbidden: The caller has no permission.'
        '404':
          description: 'Not Found: The server has not found anything matching the Request-URI.'
        '408':
          description: 'Timeout: The request timeout.'
  /dishes:
    post:
      tags:
      - dish
      summary: 'Create a new dish'
      operationId: createDish
      requestBody:
        description: 'The prepared dish request to be created.'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DishModel'
      responses:
        '201':
          description: 'The dish is created.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DishModel'
        '300':
          description: 'Moved: The resource has moved.'
        '400':
          description: 'Bad Request: The request could not be understood by the server due to malformed syntax. Something like Domain validation errors, missing data, invalid input, etc.'
        '401':
          description: 'Unauthorized: The caller is not authorized.'
        '403':
          description: 'Forbidden: The caller has no permission.'
        '404':
          description: 'Not Found: The server has not found anything matching the Request-URI.'
        '408':
          description: 'Timeout: The request timeout.'
    get:
      tags:
      - dish
      summary: 'Find all dishes'
      operationId: getAllDishes
      responses:
        '200':
          description: 'All dishes'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DishModel'
        '300':
          description: 'Moved: The resource has moved.'
        '400':
          description: 'Bad Request: The request could not be understood by the server due to malformed syntax. Something like Domain validation errors, missing data, invalid input, etc.'
        '401':
          description: 'Unauthorized: The caller is not authorized.'
        '403':
          description: 'Forbidden: The caller has no permission.'
        '404':
          description: 'Not Found: The server has not found anything matching the Request-URI.'
  /dishes/{dishId}:
    get:
      tags:
      - dish
      summary: 'Finds dish by id'
      operationId: findDishById
      parameters:
        - name: dishId
          in: path
          description: 'ID of dish to return'
          required: true
          schema:
            type: string
            maxLength: 64
            description: 'The dish identifier (UUID).'
            example: '7f7acbad-8e3b-470f-a3d4-e750360e3964'
      responses:
        '200':
          description: 'The dish'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DishModel'
        '300':
          description: 'Moved: The resource has moved.'
        '400':
          description: 'Bad Request: The request could not be understood by the server due to malformed syntax. Something like Domain validation errors, missing data, invalid input, etc.'
        '401':
          description: 'Unauthorized: The caller is not authorized.'
        '403':
          description: 'Forbidden: The caller has no permission.'
        '404':
          description: 'Not Found: The server has not found anything matching the Request-URI.'
    put:
      tags:
      - dish
      summary: 'Update an existing dish'
      operationId: updateDish
      parameters:
        - name: dishId
          in: path
          description: 'ID of dish to return'
          required: true
          schema:
            type: string
            maxLength: 64
            description: 'The dish identifier (UUID).'
            example: '7f7acbad-8e3b-470f-a3d4-e750360e3964'
      requestBody:
        description: 'The prepared age request to to be updated.'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DishModel'
      responses:
        '200':
          description: 'The dish is updated.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DishModel'
        '300':
          description: 'Moved: The resource has moved.'
        '400':
          description: 'Bad Request: The request could not be understood by the server due to malformed syntax. Something like Domain validation errors, missing data, invalid input, etc.'
        '401':
          description: 'Unauthorized: The caller is not authorized.'
        '403':
          description: 'Forbidden: The caller has no permission.'
        '404':
          description: 'Not Found: The server has not found anything matching the Request-URI.'
        '408':
          description: 'Timeout: The request timeout.'
        '409':
          description: 'Conflict: Indicates that the request could not be processed because of conflict in the current state of the resource.'
    delete:
      tags:
        - dish
      summary: 'Delete all dish by profile id'
      parameters:
        - name: dishId
          in: path
          description: 'ID of dish to return'
          required: true
          schema:
            type: string
            maxLength: 64
            description: 'The dish identifier (UUID).'
            example: '7f7acbad-8e3b-470f-a3d4-e750360e3964'
      operationId: deleteDish
      responses:
        '204':
          description: 'The dish has been sucessfully deleted.'
        '300':
          description: 'Moved: The resource has moved.'
        '400':
          description: 'Bad Request: The request could not be understood by the server due to malformed syntax. Something like Domain validation errors, missing data, invalid input, etc.'
        '401':
          description: 'Unauthorized: The caller is not authorized.'
        '403':
          description: 'Forbidden: The caller has no permission.'
        '404':
          description: 'Not Found: The server has not found anything matching the Request-URI.'
        '408':
          description: 'Timeout: The request timeout.'
...
components:
  schemas:
...
    TranslationModel:
      type: object
      required:
      - content
      - language
      properties:
        language:
          type: string
          example: 'vi'
        content:
          type: string
          example: 'content in language'
...
    AgeModel:
      type: object
      required:
      - names
      properties:
        id:
          type: string
          example: '3f3a70b3-48db-4c3d-815f-e40088811cb1'
        names:
          type: array
          items:
            $ref: '#/components/schemas/TranslationModel'
    DishModel:
      type: object
      required:
      - names
      properties:
        id:
          type: string
          example: '3f3a70b3-48db-4c3d-815f-e40088811cb1'
        names:
          type: array
          items:
            $ref: '#/components/schemas/TranslationModel'
    IngredientModel:
      type: object
      required:
      - names
      properties:
        id:
          type: string
          example: '3f3a70b3-48db-4c3d-815f-e40088811cb1'
        names:
          type: array
          items:
            $ref: '#/components/schemas/TranslationModel'
...